+++ lib/puppet/provider/elastic_plugin.rb
@@ -151,7 +151,7 @@ class Puppet::Provider::ElasticPlugin < Puppet::Provider
     env_vars['ES_JAVA_OPTS'] = env_vars['ES_JAVA_OPTS'].join(' ')
 
     env_vars.each do |env_var, value|
-      saved_vars[env_var] = ENV[env_var]
+      saved_vars[env_var] = ENV.fetch(env_var, nil)
       ENV[env_var] = value
     end