+++ lib/puppet/type/elasticsearch_template.rb
@@ -50,7 +50,7 @@ Puppet::Type.newtype(:elasticsearch_template) do
                 val['settings']['index'] = {} unless val['settings'].key? 'index'
                 (val['settings'].keys - ['index']).each do |setting|
                   new_key = if setting.start_with? 'index.'
-                              setting[6..-1]
+                              setting[6..]
                             else
                               setting
                             end
@@ -110,7 +110,7 @@ Puppet::Type.newtype(:elasticsearch_template) do
 
       fail(format('Could not find any content at %s', self[:source])) unless tmp
 
-      self[:content] = PSON.load(tmp.content)
+      self[:content] = JSON.parse(tmp.content)
     end
   end
   # rubocop:enable Style/SignalException