+++ spec/helpers/acceptance/tests/package_url_shared_examples.rb
@@ -10,8 +10,6 @@ shared_examples 'package_url acceptance tests' do |es_config|
       <<-MANIFEST
         api_timeout => 60,
         config => {
-          'cluster.name' => '#{v[:cluster_name]}',
-          'http.bind_host' => '0.0.0.0',
   #{es_config.map { |k, v| "        '#{k}' => '#{v}'," }.join("\n")}
         },
         jvm_options => [
@@ -53,11 +51,11 @@ shared_examples 'package_url acceptance tests' do |es_config|
 
     context 'via puppet paths', :with_cleanup do
       before :all do # rubocop:disable RSpec/BeforeAfterAll
-        shell "mkdir -p #{default['distmoduledir']}/another/files"
+        shell "mkdir -p #{default.puppet['codedir']}/modules/another/files"
 
         scp_to default,
                v[:elasticsearch_package][:path],
-               "#{default['distmoduledir']}/another/files/#{v[:elasticsearch_package][:filename]}"
+               "#{default.puppet['codedir']}/modules/another/files/#{v[:elasticsearch_package][:filename]}"
       end
 
       let(:manifest_class_parameters) do